refactor: nvim-magic logger should be an internal module
