+++ lua/nvim-magic/_log.lua
